First draft of returning the master token entity identity in key response data, and the provisioned entity authentication scheme. This is currently added as a sibling key/value pair in the top-level key response data. For privacy reasons this needs to change so the data is encrypted and integrity protected by the session keys in the key response data. MslContext.setEntityIdentity() has been defined to provide the entity identity when key response data is received.
Add AuthenticationUtils as a parameter to most JavaScript key exchange factories.
Validate the master token and check that the scheme is permitted in those factories.

Correct some Java KeyRequest and KeyResponse unit tests.
Add JavaScript KeyRequest and KeyResponse unit tests.

Remove JavaScript checks for !variable when parsing JSON objects as this check will return true when the value is the empty string or the numerical value 0. Only use the typeof operator to validate the data type.