Integrate fixes from clarinet.js v0.12.0, but modify by adding pause/resume functionality. This can be used to stop parsing after any element.
Change ClarinetParser.js to only use the locally modified clarinet.js and to parse data as a stream, pausing after each top-level value is extracted.

Change JsonMslTokenizer to use a single parser for the lifetime of operation instead of attempting to use multiple parsers as more data is read, to read and parse data in 64KiB chunks, and to mark/reset/skip the source input stream depending on how much data was actually parsed. This allows the source input stream to be reused. (When closed, MessageInputStream consumes all remaining payloads before closing the tokenizer.)